# docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SocialPlus Backend API",
    "description": "API a vývojové endpointy pro aplikaci SocialPlus",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
...
  },
  "paths": {
    "/api/register": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Registrace uživatele do systému",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "displayname": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "displayname",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "displayname": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "displayname"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Přihlášení uživatele do systému",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "displayname": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "displayname"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/{userId}/profile": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Získání uživatelského profilu podle ID",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "displayname": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "displayname"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/username/{username}": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Získání uživatelského profilu podle username",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "username",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "displayname": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "displayname"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/user/me": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Verifikace autentizace a získání informací o přihlášeném uživateli",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "displayname": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "displayname"
                  ]
                }
              }
            }
          },
          "4XX": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://0.0.0.0:8080",
      "description": "HTTP server"
    }
  ],
...
}